Fixing rename field mutations

+++ src/metaphor/update/mutation_delete_field.py
@@ -14,9 +14,9 @@ class DeleteFieldMutation:
     def execute(self):
         update_id = str(self.from_schema.create_update())
 
-        spec = self.from_schema.specs[self.spec_name]
-        self.from_schema.delete_field_value(self.spec_name, self.field_name)
-        self.from_schema._do_delete_field(self.spec_name, self.field_name)
+        if self.spec_name != 'root':
+            self.from_schema.delete_field_value(self.spec_name, self.field_name)
+            self.from_schema._do_delete_field(self.spec_name, self.field_name)
 
         # find and update dependent calcs
         start_agg = []
